pyproject.toml
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-optional-dependencies
[project]
name = "hcaptcha-challenger"
version = "0.11.0"
description = "🥂 Gracefully face hCaptcha challenge with MoE(ONNX) embedded solution."
license = "GPL-3.0-or-later"
authors = ["QIN2DIM <yaoqinse@gmail.com>", "Bingjie Yan <bj.yan.pa@qq.com>"]
readme = "README.md"
requires-python = ">=3.10"
homepage = "https://github.com/QIN2DIM/hcaptcha-challenger"
repository = "https://github.com/QIN2DIM/hcaptcha-challenger"
documentation = "https://github.com/QIN2DIM/hcaptcha-challenger"
keywords = ["hcaptcha", "hcaptcha-challenger", "hcaptcha-solver"]
classifiers = [
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
]
packages = [{ include = "hcaptcha_challenger", format = "sdist" }]
dependencies = [
    "bs4",
    "beautifulsoup4",
    "lxml",
    "pytz",
    "httpx[http2]",
    "loguru",
    "tqdm",
    "pyyaml",
    "importlib_metadata",
    "pydantic-settings",
    "tenacity",
    "undetected-playwright-patch",
    "playwright~=1.40.0",
    "opencv-python~=4.8.0.76",
    "pillow>=11.1.0",
    "ftfy>=6.3.1",
    "regex>=2024.11.6",
    "scipy>=1.15.2",
    "scikit-learn>=1.6.1",
    "scikit-image>=0.25.2",
    "onnxruntime>=1.21.0",
    "asyncache>=0.3.1",
    "cachetools>=5.5.2",
    "numpy<2.0.0",
    "google-genai>=1.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests", "examples"]
asyncio_mode = "auto"
filterwarnings = "ignore::DeprecationWarning"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
skip-string-normalization = true
